test(DiagonalDirectionTest): 테스트 수정
