+++ src/test/java/domain/direction/DiagonalDirectionTest.java
@@ -1,141 +1,137 @@
 package domain.direction;
 
-import static domain.direction.DiagonalDirection.DOWN_LEFT;
-import static domain.direction.DiagonalDirection.DOWN_RIGHT;
-import static domain.direction.DiagonalDirection.UP_LEFT;
-import static domain.direction.DiagonalDirection.UP_RIGHT;
+import static domain.direction.DiagonalDirection.getDirection;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 class DiagonalDirectionTest {
 
-    @Nested
-    @DisplayName("대각선 방향을 얻을 수 없다.")
-    class InvalidInput {
+    // 음수이면 위로, 양수이면 아래로 이동한 것이다.
+    int rowDifference;
+    // 음수이면 왼쪽, 양수이면 오른쪽으로 이동한 것이다.
+    int columnDifference;
 
-        @DisplayName("좌표간의 차이가 없는 경우")
-        @Test
-        void getDirection_WhenSameCoordinate() {
-            assertThatThrownBy(() -> DiagonalDirection.getDirection(0, 0))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage("이동할 수 없는 위치입니다.");
-        }
+    @DisplayName("이동이 없으면 예외가 발생한다.")
+    @Test
+    void dontMove() {
+        rowDifference = 0;
+        columnDifference = 0;
 
-        @ParameterizedTest
-        @CsvSource(value = {"1,0", "-1,0"}, delimiter = ',')
-        @DisplayName("상,하로만 이동한 경우")
-        void getDirection_WhenOnlyUpOrDown(int rowDifference, int columnDifference) {
-            assertThatThrownBy(() -> DiagonalDirection.getDirection(rowDifference, columnDifference))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage("이동할 수 없는 위치입니다.");
-        }
-
-        @ParameterizedTest
-        @CsvSource(value = {"0,1", "0,-1"}, delimiter = ',')
-        @DisplayName("좌,우로만 이동한 경우")
-        void getDirection_WhenOnlyLeftOrRight(int rowDifference, int columnDifference) {
-            assertThatThrownBy(() -> DiagonalDirection.getDirection(rowDifference, columnDifference))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage("이동할 수 없는 위치입니다.");
-        }
-
-        @ParameterizedTest
-        @CsvSource(value = {"1,2", "-3,1"}, delimiter = ',')
-        @DisplayName("상,하 방향 이동 값과 좌,우 방향 이동 값이 다른 경우")
-        void getDirection_WhenOtherCase(int rowDifference, int columnDifference) {
-            assertThatThrownBy(() -> DiagonalDirection.getDirection(rowDifference, columnDifference))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage("이동할 수 없는 위치입니다.");
-        }
+        runExceptionTest(rowDifference, columnDifference);
     }
 
-    @Nested
-    @DisplayName("오른쪽 위 대각선")
-    class UpRightDiagonal {
+    @DisplayName("위로만 이동하면 예외가 발생한다.")
+    @Test
+    void moveUp() {
+        rowDifference = -1;
+        columnDifference = 0;
 
-        Direction upRightDirection = UP_RIGHT;
-
-        @DisplayName("행과 열 값의 차이를 이용하여 방향을 얻는다.")
-        @Test
-        void getDirection() {
-            assertThat(DiagonalDirection.getDirection(-1, 1))
-                    .isEqualTo(upRightDirection);
-        }
-
-        @DisplayName("오른쪽 위 대각선 방향으로의 거리를 구한다.")
-        @Test
-        void calculateDistance() {
-            assertThat(upRightDirection.calculateDistance(-3, 3))
-                    .isEqualTo(3);
-        }
+        runExceptionTest(rowDifference, columnDifference);
     }
 
-    @Nested
-    @DisplayName("왼쪽 위 대각선")
-    class UpLeftDiagonal {
+    @DisplayName("아래로만 이동하면 예외가 발생한다.")
+    @Test
+    void moveDown() {
+        rowDifference = 1;
+        columnDifference = 0;
 
-        Direction upLeftDirection = UP_LEFT;
-
-        @DisplayName("행과 열 값의 차이를 이용하여 방향을 얻는다.")
-        @Test
-        void getDirection() {
-            assertThat(DiagonalDirection.getDirection(-1, -1))
-                    .isEqualTo(upLeftDirection);
-        }
-
-        @DisplayName("왼쪽 위 대각선 방향으로의 거리를 구한다.")
-        @Test
-        void calculateDistance() {
-            assertThat(upLeftDirection.calculateDistance(-3, -3))
-                    .isEqualTo(3);
-        }
+        runExceptionTest(rowDifference, columnDifference);
     }
 
-    @Nested
-    @DisplayName("오른쪽 아래 대각선")
-    class DownRightDiagonal {
+    @DisplayName("왼쪽으로만 이동하면 예외가 발생한다.")
+    @Test
+    void moveLeft() {
+        rowDifference = 0;
+        columnDifference = -1;
 
-        Direction downRightDirection = DOWN_RIGHT;
-
-        @DisplayName("행과 열 값의 차이를 이용하여 방향을 얻는다.")
-        @Test
-        void getDirection() {
-            assertThat(DiagonalDirection.getDirection(1, 1))
-                    .isEqualTo(downRightDirection);
-        }
-
-        @DisplayName("오른쪽 아래 대각선 방향으로의 거리를 구한다.")
-        @Test
-        void calculateDistance() {
-            assertThat(downRightDirection.calculateDistance(3, 3))
-                    .isEqualTo(3);
-        }
+        runExceptionTest(rowDifference, columnDifference);
     }
 
-    @Nested
-    @DisplayName("왼쪽 아래 대각선")
-    class DownLeftDiagonal {
+    @DisplayName("오른쪽으로만 이동하면 예외가 발생한다.")
+    @Test
+    void moveRight() {
+        rowDifference = 0;
+        columnDifference = 1;
 
-        Direction downLeftDirection = DOWN_LEFT;
+        runExceptionTest(rowDifference, columnDifference);
+    }
 
-        @DisplayName("행과 열 값의 차이를 이용하여 방향을 얻는다.")
-        @Test
-        void getDirection() {
-            assertThat(DiagonalDirection.getDirection(1, -1))
-                    .isEqualTo(downLeftDirection);
-        }
+    @DisplayName("상,하 방향으로의 이동 횟수보다 좌,우 방향으로의 이동 횟수가 많으면 예외가 발생한다.")
+    @Test
+    void moveMoreHorizonThanVertical() {
+        rowDifference = 1;
+        columnDifference = 2;
 
-        @DisplayName("오른쪽 아래 대각선 방향으로의 거리를 구한다.")
-        @Test
-        void calculateDistance() {
-            assertThat(downLeftDirection.calculateDistance(3, -3))
-                    .isEqualTo(3);
-        }
+        runExceptionTest(rowDifference, columnDifference);
+    }
+
+    @DisplayName("좌,우 방향으로의 이동 횟수보다 상,하 방향으로의 이동 횟수가 많으면 예외가 발생한다.")
+    @Test
+    void moveMoreVerticalThanHorizon() {
+        rowDifference = 2;
+        columnDifference = -1;
+
+        runExceptionTest(rowDifference, columnDifference);
+    }
+
+    @DisplayName("오른쪽 위 대각선 방향을 얻고, 이동 횟수를 계산한다.")
+    @Test
+    void calculateDistanceAfterGetUpRightDiagonal() {
+        rowDifference = -3;
+        columnDifference = 3;
+
+        DiagonalDirection actual = DiagonalDirection.getDirection(rowDifference, columnDifference);
+        DiagonalDirection expected = DiagonalDirection.UP_RIGHT;
+
+        assertThat(actual).isEqualTo(expected);
+        assertThat(actual.calculateDistance(rowDifference, columnDifference)).isEqualTo(3);
+    }
+
+    @DisplayName("왼쪽 위 대각선 방향을 얻고, 이동 횟수를 계산한다.")
+    @Test
+    void calculateDistanceAfterGetUpLeftDiagonal() {
+        rowDifference = -3;
+        columnDifference = -3;
+
+        DiagonalDirection actual = DiagonalDirection.getDirection(rowDifference, columnDifference);
+        DiagonalDirection expected = DiagonalDirection.UP_LEFT;
+
+        assertThat(actual).isEqualTo(expected);
+        assertThat(actual.calculateDistance(rowDifference, columnDifference)).isEqualTo(3);
+    }
+
+    @DisplayName("오른쪽 아래 대각선 방향을 얻고, 이동 횟수를 계산한다.")
+    @Test
+    void calculateDistanceAfterGetDownRightDiagonal() {
+        rowDifference = 3;
+        columnDifference = 3;
+
+        DiagonalDirection actual = DiagonalDirection.getDirection(rowDifference, columnDifference);
+        DiagonalDirection expected = DiagonalDirection.DOWN_RIGHT;
+
+        assertThat(actual).isEqualTo(expected);
+        assertThat(actual.calculateDistance(rowDifference, columnDifference)).isEqualTo(3);
+    }
+
+    @DisplayName("왼쪽 아래 대각선 방향을 얻고, 이동 횟수를 계산한다.")
+    @Test
+    void calculateDistanceAfterGetDownLeftDiagonal() {
+        rowDifference = 3;
+        columnDifference = -3;
+
+        DiagonalDirection actual = DiagonalDirection.getDirection(rowDifference, columnDifference);
+        DiagonalDirection expected = DiagonalDirection.DOWN_LEFT;
+
+        assertThat(actual).isEqualTo(expected);
+        assertThat(actual.calculateDistance(rowDifference, columnDifference)).isEqualTo(3);
+    }
+
+    void runExceptionTest(int rowDifference, int columnDifference) {
+        assertThatThrownBy(() -> getDirection(rowDifference, columnDifference))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("이동할 수 없는 위치입니다.");
     }
 }
